Add Quantum-Safe Algorithms to BoringSSL
We applied OQS patches to Google's BoringSSL to support:
* Quantum-safe key exchange
* Hybrid key exchange (quantum-safe + elliptic curve)
* Quantum-safe digital signatures
* Hybrid digital signatures (quantum-safe + RSA/elliptic curve)

In terms of algorithms, we added support for quantum-safe digital signature algorithms, including ML-KEM, SPHINCS-SHA2, SPHINCS-SHAKE, Falcon, CROSS, and UOV. We also incorporated quantum-safe key exchange algorithms, such as ML-KEM, BIKE, FrodoKEM, and HQC.

+++ crypto/evp/p_dh_asn1.cc
@@ -27,7 +27,7 @@ static void dh_free(EVP_PKEY *pkey) {
   pkey->pkey = NULL;
 }
 
-static int dh_size(const EVP_PKEY *pkey) {
+static size_t dh_size(const EVP_PKEY *pkey) {
   return DH_size(reinterpret_cast<const DH *>(pkey->pkey));
 }
